1. storeId int타입으로 변경

+++ src/main/java/com/example/demo/src/store/dto/StoreResponse.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 public class StoreResponse {
-    private Long storeId;
+    private int storeId;
 
     private String storeName;
